TEST: PR staging moving functions in application.js to check different scope
