+++ lib/nexmo_developer/config/webpack/environment.js
@@ -30,13 +30,10 @@ environment.plugins.prepend('Provide',
 
 new webpack.DefinePlugin({
   'process.env.ADOBE_LAUNCH_PRIMARY_CATEGORY': JSON.stringify(process.env.ADOBE_LAUNCH_PRIMARY_CATEGORY)
-});
-
-new webpack.DefinePlugin({
   'digitalData': webpack.DefinePlugin.runtimeValue(
-      function () { return JSON.stringify(dataLayer()) }, []
+      function () { return JSON.stringify(setDigitalData()) }, []
   )
-})
+});
 
 
 environment.loaders.prepend('vue', vue)
